tamil-ai-project/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
waitress==3.0.0

faster-whisper==1.0.3

python-docx==1.1.0
openpyxl==3.1.5
python-pptx==0.6.23
requests
